only take reverse complement of TCR fastq
The detectsubstitutionerrors step of DropSeq tools requires that alignments be on the sense strand, without taking the reverse complement of the biological fastq, we were getting antisense alignments. The barcode/umi fastq does not need a reverse complement to be taken. This change is non-optimal because it takes the reverse complement of both fastqs and just throws away the reverse complement of the barcode, but it requires the fewest code changes so we are going with it.